Player Attack #14 - Ranged Weapon
-Reorginistaion of scripts into different folders
-Abstraction of player data into serilizable classes for saving and loading
-Addition of ranged ability tree
-Addition of Ranged ability node
-Seperation of damage dealing methods into ranged and melee

+++ Assets/Ronan/Scripts/Player Mechanics/AbilityTree/MagicAbilityNode.cs
@@ -12,5 +12,5 @@ public class MagicAbilityNode : AbilityTreeNode
     [Header("Magic Details")]
     public MagicNodeType NodeType;
     public int HealthBonus = 0;
-    public float MagicBonus = 0f;
+    public int MagicBonus = 0;
 }
